Move Solana account types to own file
