+++ typescript/src/accounts/evm/toEvmSmartAccount.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { toEvmSmartAccount } from "./toEvmSmartAccount.js";
-import { EvmAccount } from "../types.js";
+import { EvmAccount } from "./types.js";
 import { Address } from "../../types/misc.js";
 import {
   CdpOpenApiClientType,
