+++ typescript/src/index.ts
@@ -1,2 +1,2 @@
 export { CdpClient } from "./client/cdp.js";
-export type { EvmServerAccount, EvmSmartAccount } from "./accounts/types.js";
+export type { EvmServerAccount, EvmSmartAccount } from "./accounts/evm/types.js";
